apps/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Python agent app for LinguaBoost"
authors = [
    {name = "LinguaBoost Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "asyncio-mqtt>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
